feat: add simple test, free text properly, fix len
add test.c with simple way to test ft_printf against system printf.
improve Makefile to be able to build and run tests.
add ft_vprintf, which was needed to be able to run tests properly.
change type of in-between text to 't', so we can later free it properly.
move check for flags to fill_data so it gets checked at the end.
load data from union as the size it was entered as in nbrlen.
implement the '0' flag properly for numbers.